refactor: move session routes to separate module
