+++ src/routers/__init__.py
@@ -1,0 +1,2 @@
+from src.routers.sessions import session_router
+
